Move the CSVBasedLicenseKnowledgeBase to core-workflow-steps
