+++ core/core-workflow-steps/src/main/java/org/eclipse/sw360/antenna/knowledgebase/CSVBasedLicenseKnowledgeBase.java
@@ -10,7 +10,7 @@
  * SPDX-License-Identifier: EPL-2.0
  */
 
-package org.eclipse.sw360.antenna.knowledgebase.csv;
+package org.eclipse.sw360.antenna.knowledgebase;
 
 import org.apache.commons.csv.CSVFormat;
 import org.apache.commons.csv.CSVParser;
