+++ core/core-workflow-steps/src/test/resources/licenses/AFL-1.1.txt
@@ -1,0 +1,1 @@
+Should be license text of AFL-1.1.
